+++ codaamok.build/Files/custom.build.ps1
@@ -1,0 +1,10 @@
+[CmdletBinding()]
+param ()
+
+task Pre {
+
+}
+
+task Post {
+    
+}